Convert everything to typescript and split files
