+++ migration/2020Q3_to_2020Q4/migration/decrypt.ts
@@ -1,15 +1,13 @@
-"use strict";
-
 const { exec } = require("child_process");
 
 const phpScriptPath = "migration/decrypt.php";
-module.exports = (encryptedStrings, key, defaultIv) => {
+export default (encryptedStrings: string[], key: string, defaultIv: any) => {
   return new Promise((resolve, reject) => {
     exec(
       `php ${phpScriptPath} '${JSON.stringify(
         encryptedStrings
       )}' '${JSON.stringify(key)}' '${JSON.stringify(defaultIv)}'`,
-      (err, res) => {
+      (err: any, res: string) => {
         if (err) reject(err);
         resolve(JSON.parse(res));
       }
