--- migration/2020Q3_to_2020Q4/migration/start.js
@@ -1,3 +1,0 @@
-import migration from "./migrate.js";
-
-migration();
