+++ migration/2020Q3_to_2020Q4/migration/store.ts
@@ -1,0 +1,35 @@
+import config from "config";
+import cassandra from "cassandra-driver";
+import DatabaseServiceClass from "./services/db/index";
+
+const configuration: any = {
+  db: config.get("db"),
+  encryption: config.get("encryption"),
+};
+
+class Store {
+  private databaseService: DatabaseServiceClass = null;
+  private cassandraClient: any = null;
+
+  async initOrmClient() {
+    this.databaseService = new DatabaseServiceClass("cassandra", {
+      ...config.get("db"),
+    });
+  }
+  getOrmClient() {
+    return this.databaseService;
+  }
+
+  async initCassandraClient() {
+    this.cassandraClient = new cassandra.Client({
+      contactPoints: configuration.db.contactPoints,
+      localDataCenter: configuration.db.localDataCenter,
+      keyspace: configuration.db.keyspace,
+    });
+  }
+  getCassandraClient() {
+    return this.cassandraClient;
+  }
+}
+
+export default new Store();
